ENH: allow the terms of the cost function to be weighted by user supplied weights ENH: Point type changed to be double ENH: make mean/pca image type a template parameter

+++ Code/Algorithms/itkShapePriorMAPCostFunctionBase.h
@@ -87,7 +87,7 @@ public:
   typedef typename NodeContainerType::ConstPointer  NodeContainerPointer;
 
   /** Type of the shape signed distance function. */
-  typedef ShapeSignedDistanceFunction<float,
+  typedef ShapeSignedDistanceFunction<double,
                                       itkGetStaticConstMacro(ImageDimension)> ShapeFunctionType;
   typedef typename ShapeFunctionType::Pointer            ShapeFunctionPointer;
